CUI-5 move helper files (graph.py tallytable.py testdb.py) to utils folder and edit views.py accordingly
